Manage server-side API changes + cleanup
